LangChain/requirements.txt
fastapi
langchain-openai
langchain-community
python-dotenv
psycopg2-binary
uvicorn